entry edit

+++ screen/EditEntry.js
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { View, Text } from "react-native";
+import { View, Text, Alert } from "react-native";
 import ButtonPressable from "../component/ButtonPressable";
 import { deleteFromDB, updateDB } from "../dataSource/FireStoreHelper";
 
@@ -8,14 +8,46 @@ export default function EditEntry({ route, navigation }) {
     const { calory, name, id, reviewed } = route.params.meal;
 
     const onDel = function d(id) {
-        deleteFromDB({docid: id});
-        navigation.popToTop();
-    }
+        Alert.alert(
+            'Confirmation of delete',
+            'Are you sure you want to delete this entry?',
+            [
+              {
+                text: 'No',
+                style: 'cancel',
+              },
+              {
+                text: 'Yes',
+                onPress: () => {
+                    deleteFromDB({docid: id});
+                    navigation.popToTop();
+                },
+              },
+            ],
+            { cancelable: false }
+        );
+    };
 
     const onUd = function u(id) {
-        updateDB({docid: id});
-        navigation.popToTop();
-    }
+        Alert.alert(
+            'Confirmation of review',
+            'Are you sure you want to review this entry?',
+            [
+              {
+                text: 'No',
+                style: 'cancel',
+              },
+              {
+                text: 'Yes',
+                onPress: () => {
+                    updateDB({docid: id});
+                    navigation.popToTop();
+                },
+              },
+            ],
+            { cancelable: false }
+        );
+    };
 
     return (
         <View>
@@ -43,13 +75,15 @@ export default function EditEntry({ route, navigation }) {
                         del
                     </Text>
                 </ButtonPressable>
-                <ButtonPressable
-                    onPressed={() => {onUd(id);}}
-                >
+                {!reviewed && (
+                    <ButtonPressable
+                        onPressed={() => {onUd(id);}}
+                    >
                     <Text>
                         rev
                     </Text>
-                </ButtonPressable>
+                    </ButtonPressable>
+                )}
             </View>
         </View>
     );
